+++ globals/jira_globals.py
