+++ tests/test_login_to_jira.py
@@ -1,6 +1,6 @@
 import pytest
 import requests
-from tests.jira_globals import *
+from globals.jira_globals import *
 from tests.fixtures.json_fixtures import JsonFixtures
